move TestPipelineVersions
