+++ src/server/pachyderm_test.go
@@ -44,7 +44,6 @@ import (
 	ppspretty "github.com/pachyderm/pachyderm/v2/src/server/pps/pretty"
 	"github.com/pachyderm/pachyderm/v2/src/storage"
 
-	"github.com/pachyderm/pachyderm/v2/src/internal/ancestry"
 	"github.com/pachyderm/pachyderm/v2/src/internal/backoff"
 	"github.com/pachyderm/pachyderm/v2/src/internal/cmdutil"
 	"github.com/pachyderm/pachyderm/v2/src/internal/errors"
@@ -8409,41 +8408,6 @@ func TestInspectJob(t *testing.T) {
 	_, err = c.InspectJob(pfs.DefaultProjectName, repo, ci.Commit.Id, false)
 	require.YesError(t, err)
 	require.True(t, errutil.IsNotFoundError(err))
-}
-
-func TestPipelineVersions(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration tests in short mode")
-	}
-
-	t.Parallel()
-	c, _ := minikubetestenv.AcquireCluster(t)
-
-	dataRepo := tu.UniqueString("TestPipelineVersions_data")
-	require.NoError(t, c.CreateRepo(pfs.DefaultProjectName, dataRepo))
-
-	pipeline := tu.UniqueString("TestPipelineVersions")
-	nVersions := 5
-	for i := 0; i < nVersions; i++ {
-		require.NoError(t, c.CreatePipeline(pfs.DefaultProjectName,
-			pipeline,
-			"",
-			[]string{fmt.Sprintf("%d", i)}, // an obviously illegal command, but the pipeline will never run
-			nil,
-			&pps.ParallelismSpec{
-				Constant: 1,
-			},
-			client.NewPFSInput(pfs.DefaultProjectName, dataRepo, "/*"),
-			"",
-			i != 0,
-		))
-	}
-
-	for i := 0; i < nVersions; i++ {
-		pi, err := c.InspectPipeline(pfs.DefaultProjectName, ancestry.Add(pipeline, nVersions-1-i), true)
-		require.NoError(t, err)
-		require.Equal(t, fmt.Sprintf("%d", i), pi.Details.Transform.Cmd[0])
-	}
 }
 
 func TestDeferredProcessing(t *testing.T) {
